+++ cider-repl.el
@@ -1061,7 +1061,8 @@ constructs."
   (add-to-list 'completion-at-point-functions
                'cider-complete-at-point)
   (set-syntax-table cider-repl-mode-syntax-table)
-  (cider-turn-on-eldoc-mode)
+  (cider-eldoc-setup)
+  (eldoc-mode +1)
   ;; At the REPL, we define beginning-of-defun and end-of-defun to be
   ;; the start of the previous prompt or next prompt respectively.
   ;; Notice the interplay with `cider-repl-beginning-of-defun'.